This enables using alternative githubs on file.
Because the current file format is supposed to allow the standard format
as well as something slightly different it is mutating the passed in
config in order to work. This currently missed off the baseUrl option
which allows us to contact different githubs. This is a bit of a hack
but I don't want to break backwards compatibility with this format.