Change Fourth Wall to be a Django 1.7.1 project
This commit uses the [Heroku Django template][1] to create a Django 1.7.1
project for Fourth Wall.

At the moment this is just the existing code moved around to fit in a
Django application. My intention is to move this away from being an
entirely client-side application into something where configuration
is stored in a database and the application is hosted using a service
like Heroku.

Lots of people who have used this project have struggled with using Gists
for configuration, because of GitHub's model where multiple people can't
edit a single Gist.

I've removed the tests because they're using a version of Jasmine that
was released in December 2012 and I think there are better ways of doing
things now.

[1]: https://github.com/heroku/heroku-django-template