Use python-social-auth to login with GitHub
This commit adds authentication with GitHub and GitHub Enterprise
using python-social-auth.

The flow isn't perfect yet but I'm pretty happy with it:

- You must sign in with GitHub (public) first
- You can only sign in to GitHub Enterprise if you're signed in to
  GitHub already
- If you're signed in to GitHub and you sign in to GitHub Enterprise,
  your 2 tokens are associated with the same account in the database.

Tokens are returned in the list of repositories, and the JavaScript
has been updated to use those tokens rather than try and pry one out
of the query string.