Fetch more teams when looking up team slug to ID
Looking up a team via the GH API is done by ID, not slug, so we
have to lookup the list of teams and manually map the slug to an
ID for further API calls.

However, the team list API response is paginated, and if the team
you're trying to map is not in the first page of results then we
won't be able to map to the ID.

Increase the number of teams returned in a page of results, by
default this is 30, but can be up to 100 [1]. We have 65 teams
at the moment, so 100 will work for now and into the near future
but we may run into the same problem again in the future. If so
we could look at walking through the paginated results, but that
felt like overkill for now.

[1]: https://developer.github.com/v3/#pagination