Allow filtering by user to be disabled by param
Currently you can set a list of users in the Gist configuration
and only show PRs owned by those users. This is generally used to
limit PRs to members of a single team where they share repos with
other teams and want to focus on their changes.

To switch between filtering and not you need to edit the gist,
which more involved than a query param, especially as you normally
want the same repo list / styling / etc.

I'm also planning to add the filter-by-user behaviour to
team-based config, and as a default limiting PRs to members of
that team. This param gives an easy way to bring back the previous
behaviour.