Make filtering by user opt-in by param (+ for teams)
Currently you can set a list of users in the Gist configuration
and only show PRs owned by those users. This is generally used to
limit PRs to members of a single team where they share repos with
other teams and want to focus on their changes.

To switch between filtering and not you need to edit the gist,
which is more involved than a query param, especially as you
normally want the same repo list / styling / etc.

Prevoiusly if users were set in a gist then they would
automatically be used to filter PRs, however, after discussion
with Edds he convinced me that it's a bad default, we should
encourage being aware of other teams/contributors work on repos.

This doesn't remove the code behaviour, just changes the default
to promote better human behaviour. As this also allows you to
filter by users for team-based config, and having filtering on by
default would be even worse.

Initialise `FourthWall.importantUsers` as it may not be set by
either team config or gist config, as they can be used at the
same time.

Note: this makes the call to get team members even if filtering is
not used. This is it would be useful to use `importantUsers` to
highlight important PRs, without entirely ommiting others.

Note: Urgh, javascript array merging.