Team-based config only showing PRs by team members
This mirrors including `users.json` in gist-based config, but is
implicit from the GH team itself.

This will change the default behaviour, where team-based config
showed all PRs you'll now need to set the `showallusers` query
param to see PRs by non-team members.

I'd rather have not changed the default, but I want a param that
controls users set in JSON config, and from teams. It makes sense
that if gist-config sets users then they should be used without
needing to opt in, and I think teams are more widely using that
config vs team-based, so it's less disruptive to require those
using team-based config to add the parameter.

If this is contentious, then i'd be will to switch to opting-in
to user filtering instead, but i'd like to be able to get
team-based filtering in somehow.