Fetch as many repos for a team as possible
By default the github API returns 30 items for a paginated set of
results. The repo list is paginated and on a team with more than
30 repos not all repos were shown

GH lets you request larger page sizes, up to 100, which won't
solve the problem forever, but makes it less likely we'll miss
repos when using team-based config.