Refactor fetchDefer to allow AJAX params as a hash
Team and repo fetching need to set query params to control
pagination, which had to be done by building the params in to the
URL being requested through `fetchDefer`.

jQuery has a `data` for setting AJAX params via a hash, avoiding
string construction which is error prone and makes intent less
clear to someone reading the code.

Update `fetchDefer` to allow the `data` option and update the
calls which use query params to use the new option.