Add support for marking repositories as important
This is for use in conjunction with the "filterusers" parameter, to make
PRs on some repositories always be shown, even if not opened by one of
the specified users.

This allows us to have a wall which lists any PRs on repositories
"owned" by our team, together with any PRs on other repositories made by
team members.

Implemented by adding a `unimportant-repo` class to pull requests from
any repo which isn't marked as important, and only hiding pull requests
which are from both unimportant repos and users.