Allow repositories to be ignored
This change adds a new file type to the gist that allows a list of
repositories to be provided that will be ignored by fourth wall. This is
useful if you load your repositories by team name but want to prevent
some being accessed. For example, if they contain sensitive information
that you do not accessible from a machine sitting unattended in an
office.