Add ability to ignore WIP commits
We have met the need, to ignore the WIP commits, in order to focus on
actual PR that may be important.

This change, allows us to add an extra GET parameter, to the URL which
then will be taken into consideration to hide WIP PRs.

The usage, has also been documented.