Check individual users comments for their status
If a user requests changes to a PR they may then later give it approval
once the changes have been made.

Previously this would not change the status displayed for the PR, as the
code was just checking if ANY comment had the state of
'CHANGES_REQUESTED' regardless of when, or if it was followed by an
'ACCEPTED' state.

This PR changes the code to look at a users comments in order and check
if the state of the last commit that was either 'ACCEPTED' or
'CHANGES_REQUESTED'. It does this or all users.