Monitor default branch instead of master
For some repos, the default branch is set to something other than
"master". This updates the master status tracking functionality to track
whatever the default branch is set to in Github instead of being
hardcoded to "master".

This is only supported when fetching repos by team at the moment as
that's currently the only part of the code that fetches the repo
information from Github - the gist and file methods get all the repo
data they need from the JSON data in the files.