Add http-server to dev dependencies
We need to serve index.html and index.js for tests, use the simple
Node.js static file http server to do this [http-server] to do this.

This commit configures jest-puppeteer to run the server with jest. This
acts as a replacement for `python -m SimpleHTTPServer` in tests.

This means:

- You can now run tests with `npm test` and have the server launch and
stop automatically.

- You no longer need to run `python -m SimpleHTTPServer` before running
`npm test`.

[http-server]: https://github.com/http-party/http-server#readme