Warn users about retirement on 22 September
We recently discovered that the Digital Marketplace team was responsible for this repo. We don't have the time or expertise to fix the accessibility issues, and we're moving to CCS soon. We asked, and no other team in GDS/CDIO wanted to take it on. So we're going to retire it in an orderly fashion.

3 weeks is probably long enough for people to notice. They'll be able to find something else, or let us know if they want to take it on to stop the retirement. Otherwise, we'll switch it off after that date.